ai_starter/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-starter"
version = "1.0.0"
description = "Enterprise-grade AI agent template with RAG, MCP, LangChain/LangGraph support"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "AI Starter Team"}
]
keywords = ["ai", "agent", "llm", "rag", "mcp", "langchain", "langgraph", "ollama", "autonomous"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0",
    "httpx>=0.25.0",
    "structlog>=23.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
langchain = [
    "langchain>=0.1.0",
    "langgraph>=0.0.1",
]
all = [
    "ai-starter[dev,langchain]",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai_starter"
Documentation = "https://github.com/yourusername/ai_starter#readme"
Repository = "https://github.com/yourusername/ai_starter"
Issues = "https://github.com/yourusername/ai_starter/issues"

[project.scripts]
ai-starter = "ai_starter.main:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_starter*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long
